% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEST4EXP.R
\name{TEST4EXP}
\alias{TEST4EXP}
\title{Gives test statistic value and critical value for one parameter exponential distribution with group size 4}
\usage{
TEST4EXP(x1, x2, x3, x4, alpha)
}
\arguments{
\item{x1}{is sample from Exp Population 1.}

\item{x2}{is sample from Exp Population 2.}

\item{x3}{is sample from Exp Population 3.}

\item{x4}{is sample from Exp Population 4.}

\item{alpha}{is numeric value (level of significance)}
}
\description{
Gives test statistic value and critical value for one parameter exponential distribution with group size 4
}
